requirements.txt
appdirs==1.4.4
Babel==2.9.1
black==21.7b0
click==8.0.1
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
pip-autoremove==0.9.1
py-money==0.5.0
pytz==2021.1
regex==2021.8.3
tomli==1.2.1
Werkzeug==2.0.1
